requirements.txt
alembic==1.8.0
aniso8601==9.0.1
asttokens==2.0.5
async-timeout==4.0.2
Babel==2.10.1
backcall==0.2.0
bidict==0.22.0
blinker==1.4
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
colorclass==2.2.2
commonmark==0.9.1
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.1
docopt==0.6.2
dominate==2.6.0
elastic-transport==8.1.2
elasticsearch==8.2.2
email-validator==1.2.1
executing==0.8.3
Flask==2.1.2
Flask-Babel==2.0.0
Flask-BabelEx==0.9.4
Flask-HTTPAuth==4.7.0
Flask-Limiter==2.4.5.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
ghostscript==0.7
greenlet==1.1.2
guess-language-spirit==0.5.3
gunicorn==20.1.0
idna==3.3
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
limits==2.6.2
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
numpy==1.22.4
packaging==21.3
parso==0.8.3
passlib==1.7.4
phonenumbers==8.12.49
pickleshare==0.7.5
Pillow==9.1.1
pip-upgrader==1.4.15
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pure-eval==0.2.2
Pygments==2.12.0
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
pytz==2022.1
redis==4.3.3
requests==2.27.1
rich==12.4.4
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.37
stack-data==0.2.0
terminaltables==3.1.10
traitlets==5.2.2
typing_extensions==4.2.0
urllib3==1.26.9
validators==0.20.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
WTForms-Components==0.10.5
